User directory: Config in Client Settings and load fonts from user dir
