+++ megamek/i18n/megamek/client/messages.properties
@@ -1214,6 +1214,7 @@ CommonSettingsDialog.locale.Russian=\u0420\u0443\u0441\u0441\u043a\u0438\u0439
 CommonSettingsDialog.locale.Spanish=Español
 CommonSettingsDialog.locale=Language/Sprache/\u042f\u0437\u044b\u043a:
 CommonSettingsDialog.logFileName=Game log filename:
+CommonSettingsDialog.userDir=User Files Directory:
 CommonSettingsDialog.main=Main
 CommonSettingsDialog.audio=Audio
 CommonSettingsDialog.miniMap=Mini Map
