+++ megamek/src/megamek/client/ui/swing/util/FontHandler.java
@@ -19,6 +19,7 @@
 package megamek.client.ui.swing.util;
 
 import megamek.MMConstants;
+import megamek.common.preference.PreferenceManager;
 import org.apache.logging.log4j.LogManager;
 
 import java.awt.*;
@@ -78,7 +79,14 @@ public final class FontHandler {
     }
 
     private void initializeFonts() {
+        LogManager.getLogger().info("Loading fonts from " + MMConstants.FONT_DIRECTORY);
         parseFontsInDirectory(new File(MMConstants.FONT_DIRECTORY));
+
+        String userDir = PreferenceManager.getClientPreferences().getUserDir();
+        LogManager.getLogger().info("Loading fonts from " + userDir);
+        parseFontsInDirectory(userDir);
+
+        LogManager.getLogger().info("Loading fonts from Java's GraphicsEnvironment");
         for (String fontName : GraphicsEnvironment.getLocalGraphicsEnvironment().getAvailableFontFamilyNames()) {
             allFontNames.add(fontName);
             Font font = Font.decode(fontName);
@@ -95,9 +103,20 @@ public final class FontHandler {
      *
      * @param directory the directory to parse
      */
+    public static void parseFontsInDirectory(String directory) {
+        parseFontsInDirectory(new File(directory));
+    }
+
+    /**
+     * Searches the provided directory and all subdirectories and registers any truetype
+     * fonts from .ttf files it finds.
+     *
+     * @param directory the directory to parse
+     */
     public static void parseFontsInDirectory(final File directory) {
         final String[] filenames = directory.list();
         if (filenames == null) {
+            LogManager.getLogger().warn(directory + " is not a directory!");
             return;
         }
